refac: 테스트에 적합하게 리팩토링
